requirements.txt
rust
pyyaml
fastapi
uvicorn
docker
pydantic